third_party/blink/web_tests/external/wpt/css/css-contain/container-queries/canvas-as-container-006-expected.txt

This is a testharness.js-based test.
PASS Initially display:none, not focusable
FAIL Focusable after container size change assert_equals: expected Element node <div id="target" tabindex="1"></div> but got Element node <body><p>Test passes if there is a filled green square.</...
Harness: the test ran to completion.
